+++ src/ptpapi/scripts/ptp_reseed.py
@@ -435,7 +435,7 @@ def process(cli_args):
     loaded_paths = []
 
     if args.client:
-        if args.client[0].startswitsh("file://"):
+        if args.client[0].startswith("file://"):
             client = args.client[0]
         else:
             client = libtc.parse_libtc_url(args.client[0])